+++ spec/intridea/changelog_line_parser_spec.rb
@@ -5,6 +5,7 @@ describe Danger::Changelog::ChangelogLineParser do
     context 'changelog entry line' do
       context 'when valid' do
         let(:line) { '* [#123](https://github.com/dblock/danger-changelog/pull/123): Test - [@dblock](https://github.com/dblock).' }
+
         it 'returns ChangelogEntryLine' do
           expect(described_class.parse(line)).to be_a(Danger::Changelog::ChangelogEntryLine)
         end
@@ -26,6 +27,7 @@ describe Danger::Changelog::ChangelogLineParser do
 
     context 'changelog your contribution here line' do
       let(:line) { "* Your contribution here.\n" }
+
       it 'returns ChangelogPlaceholderLine' do
         expect(described_class.parse(line)).to be_a(Danger::Changelog::ChangelogPlaceholderLine)
       end
@@ -34,6 +36,7 @@ describe Danger::Changelog::ChangelogLineParser do
     context 'changelog header line' do
       context 'when one hash' do
         let(:line) { '# Version 1.1.1' }
+
         it 'returns ChangelogHeaderLine' do
           expect(described_class.parse(line)).to be_a(Danger::Changelog::ChangelogHeaderLine)
         end
@@ -41,6 +44,7 @@ describe Danger::Changelog::ChangelogLineParser do
 
       context 'when two hashes' do
         let(:line) { '## Version 1.1.1' }
+
         it 'returns ChangelogHeaderLine' do
           expect(described_class.parse(line)).to be_a(Danger::Changelog::ChangelogHeaderLine)
         end
@@ -48,6 +52,7 @@ describe Danger::Changelog::ChangelogLineParser do
 
       context 'when three hashes' do
         let(:line) { '### Version 1.1.1' }
+
         it 'returns ChangelogHeaderLine' do
           expect(described_class.parse(line)).to be_a(Danger::Changelog::ChangelogHeaderLine)
         end